+++ database/migrations/2020_02_03_172248_create_transactions_table.php
@@ -15,7 +15,7 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('kuantitas');
+            $table->double('kuantitas');
             $table->double('total_harga');
             $table->bigInteger('no_urut');
